2017 Beta 2 Changes

-Improved error handling and easier reporting
-Autonomy bugfixes
-Subscriptions are now logged to console
-Main GUI layout improvements
-Control scheme selection is more readable
-GPS Start Location setting
-Improved Camera Board GUI and bugfix
-2017 Science Updates
	-2017 Sensor support
	-Telemetry bugfixes
	-Science Arm support
	-2017 Spectromemter support
-2017 Arm Updates
	-Power controls
	-Fault messages
	-Textbox bugfixes
	-Current sensing
	-Non-legacy dataids
	-Arm not stopping bugfix
-Gimbal zooming on controller
-Gimbal 2 input support
-Drive non-legacy dataids
-BMS control of buzzer and fan
-Cleaner InputMapping file structure
-Lots of internal code clean-up

+++ RED/Models/Modules/AutonomyModel.cs
@@ -2,7 +2,7 @@
 {
     internal class AutonomyModel
     {
-        internal float latitudeInput;
-        internal float longitudeInput;
+        internal double latitudeInput;
+        internal double longitudeInput;
     }
 }
